backend/requirements.txt
# ==============================================
# Python Dependencies for Contact Form Backend
# ==============================================
# Install with: pip install -r requirements.txt
# ==============================================

fastapi
uvicorn[standard]
python-dotenv
pydantic
